Move dagger module classes to respective shared module.
